requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
drf-yasg
sqlparse
psycopg2-binary
python-dotenv
Pillow